Crate wafel_data_type

Representation of Wafel data types and values, mostly corresponding to C data types.

Modules

shallow
Non-recursive version of data types that can be resolved into regular data types.

Structs

Address
A raw pointer value that can be stored in memory.

Field
A field in a struct or union.

NotAnArrayOrPointerError

TypeName
A symbolic reference to a type definition, e.g. struct Foo.

ValueTypeError

Enums

DataType
A representation of a C data type.

FloatType
Float types of different sizes.

IntType
Integer types of different sizes and signedness.

Namespace
The C type namespaces.

Value
A dynamically typed value.

Type Definitions

DataTypeRef
A reference to a DataType.

FloatValue
An integer value.

IntValue
An integer value.