wafel_data_type
Enum DataType

pub enum DataType {
    Void,
    Int(IntType),
    Float(FloatType),
    Pointer {
        base: DataTypeRef,
        stride: Option<usize>,
    },
    Array {
        base: DataTypeRef,
        length: Option<usize>,
        stride: usize,
    },
    Struct {
        fields: IndexMap<String, Field>,
    },
    Union {
        fields: IndexMap<String, Field>,
    },
    Name(TypeName),
}

A representation of a C data type.

Variants

Void

Void, typically used as a pointer target or function return type.

Int(IntType)

An integer type.

Float(FloatType)

A float type.

Pointer

A pointer type.

Fields

base: DataTypeRef

The type being pointed to.

stride: Option<usize>

The size of the type being pointed to, if known.

This is used when applying a subscript to a pointer, e.g. p[3]. It is otherwise not necessary.

Array

An array type, optionally with a length.

Fields

base: DataTypeRef

The element type.

length: Option<usize>

The length of the array, if known.

stride: usize

The size of the element type.

This is used when indexing into the array.

Struct

A struct type.

Fields

fields: IndexMap<String, Field>

The fields contained in the struct.

Anonymous fields should be given a name on construction, typically __anon.

Union

A union type.

Fields

fields: IndexMap<String, Field>

The fields contained in the union.

Anonymous fields should be given a name on construction, typically __anon.

Name(TypeName)

A symbolic reference to a type definition, e.g. struct Foo.

Implementations

impl DataType

pub fn is_void(&self) -> bool

Return true if the data type is void.

pub fn try_as_void(&self) -> Result<(), DataTypeError>

Return an error if the data type is not void.

pub fn is_int(&self) -> bool

Return true if the data type is an integer type.

pub fn try_as_int(&self) -> Result<IntType, DataTypeError>

Convert the data type to an int type.

pub fn is_float(&self) -> bool

Return true if the data type is a float type.

pub fn try_as_float(&self) -> Result<FloatType, DataTypeError>

Convert the data type to a float type.

pub fn is_pointer(&self) -> bool

Return true if the data type is a pointer type.

pub fn try_as_pointer( &self, ) -> Result<(&DataTypeRef, Option<usize>), DataTypeError>

Convert the data type to a pointer type.

pub fn is_array(&self) -> bool

Return true if the data type is an array type.

pub fn try_as_array( &self, ) -> Result<(&DataTypeRef, Option<usize>, usize), DataTypeError>

Convert the data type to an array type.

pub fn is_struct(&self) -> bool

Return true if the data type is a struct type.

pub fn try_as_struct(&self) -> Result<&IndexMap<String, Field>, DataTypeError>

Convert the data type to a struct type.

pub fn is_union(&self) -> bool

Return true if the data type is a union type.

pub fn try_as_union(&self) -> Result<&IndexMap<String, Field>, DataTypeError>

Convert the data type to a union type.

pub fn is_name(&self) -> bool

Return true if the data type is a type name.

pub fn try_as_name(&self) -> Result<&TypeName, DataTypeError>

Convert the data type to a type name.

pub fn stride(&self) -> Result<Option<usize>, DataTypeError>

Return the stride for an array or pointer type.

pub fn struct_field(&self, name: &str) -> Result<&Field, DataTypeError>

Look up a field by name in a struct type.