Enum wafel_data_type::DataType

pub enum DataType {
    Void,
    Int(IntType),
    Float(FloatType),
    Pointer {
        base: DataTypeRef,
        stride: Option<usize>,
    },
    Array {
        base: DataTypeRef,
        length: Option<usize>,
        stride: usize,
    },
    Struct {
        fields: IndexMap<String, Field>,
    },
    Union {
        fields: IndexMap<String, Field>,
    },
    Name(TypeName),
}

A representation of a C data type.

Variants

Void

Void, typically used as a pointer target or function return type.

Int(IntType)

An integer type.

Float(FloatType)

A float type.

Pointer

Fields

base: DataTypeRef

The type being pointed to.

stride: Option<usize>

The size of the type being pointed to, if known.

This is used when applying a subscript to a pointer, e.g. p[3]. It is otherwise not necessary.

A pointer type.

Array

Fields

base: DataTypeRef

The element type.

length: Option<usize>

The length of the array, if known.

stride: usize

The size of the element type.

This is used when indexing into the array.

An array type, optionally with a length.

Struct

Fields

fields: IndexMap<String, Field>

The fields contained in the struct.

Anonymous fields should be given a name on construction, typically __anon.

A struct type.

Union

Fields

fields: IndexMap<String, Field>

The fields contained in the union.

Anonymous fields should be given a name on construction, typically __anon.

A union type.

Name(TypeName)

A symbolic reference to a type definition, e.g. struct Foo.

Implementations

impl DataType

pub fn is_void(&self) -> bool

Return true if the data type is void.

pub fn is_int(&self) -> bool

Return true if the data type is an integer type.

pub fn is_float(&self) -> bool

Return true if the data type is a float type.

pub fn is_pointer(&self) -> bool

Return true if the data type is a pointer type.

pub fn is_array(&self) -> bool

Return true if the data type is an array type.

pub fn is_struct(&self) -> bool

Return true if the data type is a struct type.

pub fn is_union(&self) -> bool

Return true if the data type is a union type.

pub fn is_name(&self) -> bool

Return true if the data type is a type name.

pub fn stride(&self) -> Result<Option<usize>, NotAnArrayOrPointerError>

Return the stride for an array or pointer type.