Enum wafel_data_type::Value

pub enum Value {
    None,
    Int(IntValue),
    Float(FloatValue),
    String(String),
    Address(Address),
    Struct(Box<IndexMap<String, Value>>),
    Array(Vec<Value>),
}

A dynamically typed value.

Variants

None

Represents the lack of a value.

For example, when evaluating a data path and ? is used on a null pointer, Value::None is returned for the entire path.

Int(IntValue)

An integer value, regardless of the underlying IntType size.

Float(FloatValue)

A float value, regardless of the underlying FloatType size.

String(String)

A string value.

Address(Address)

An address value.

Struct(Box<IndexMap<String, Value>>)

A struct value.

If a field’s name is present in the original struct definition, it will match the name used here. Anonymous fields will be given a name, typically __anon.

Array(Vec<Value>)

An array value.

Implementations

impl Value

pub fn is_none(&self) -> bool

Return true if the value is Value::None.

pub fn as_none(&self)

Panic if the value is not Value::None.

pub fn try_as_none(&self) -> Result<(), ValueTypeError>

Return an error if the value is not Value::None.

pub fn option(&self) -> Option<&Value>

Return None if the value is none, and Some(self) otherwise.

pub fn as_int(&self) -> IntValue

Convert the value to an int, panicking if it is not an int.

pub fn try_as_int(&self) -> Result<IntValue, ValueTypeError>

Convert the value to an int.

pub fn as_int_lenient(&self) -> IntValue

Convert the value to an int, allowing in-range floats that are integers.

Panics if the conversion fails.

pub fn try_as_int_lenient(&self) -> Result<IntValue, ValueTypeError>

Convert the value to an int, allowing in-range floats that are integers.

pub fn as_usize(&self) -> usize

Convert the value to a usize, panicking on failure.

pub fn try_as_usize(&self) -> Result<usize, ValueTypeError>

Convert the value to a usize.

pub fn as_float(&self) -> FloatValue

Convert the value to a float, panicking if the value is not a float.

pub fn try_as_float(&self) -> Result<FloatValue, ValueTypeError>

Convert the value to a float.

pub fn as_float_lenient(&self) -> FloatValue

Convert the value to a usize, allowing in-range integers.

Panics on failure.

pub fn try_as_float_lenient(&self) -> Result<FloatValue, ValueTypeError>

Convert the value to a float, allowing in-range integers.

pub fn as_f32(&self) -> f32

Convert the value to a float, and then truncate to an f32, panicking on failure.

pub fn try_as_f32(&self) -> Result<f32, ValueTypeError>

Convert the value to a float and then truncate to an f32.

pub fn as_str(&self) -> &str

Convert the value to a string, panicking on failure.

pub fn try_as_str(&self) -> Result<&str, ValueTypeError>

Convert the value to a string.

pub fn as_address(&self) -> Address

Convert the value to an address, panicking on failure.

pub fn try_as_address(&self) -> Result<Address, ValueTypeError>

Convert the value to an address.

pub fn as_struct(&self) -> &IndexMap<String, Value>

Convert the value to a struct and return its fields, panicking on failure.

pub fn try_as_struct(&self) -> Result<&IndexMap<String, Value>, ValueTypeError>

Convert the value to a struct and return its fields.

pub fn as_array(&self) -> &[Value]

Convert the value to an array and return its elements, panicking on failure.

pub fn try_as_array(&self) -> Result<&[Value], ValueTypeError>

Convert the value to an array and return its elements.

pub fn as_array_with_len(&self, length: usize) -> &[Value]

Convert the value to an array and return its elements, panicking on failure.

pub fn try_as_array_with_len(
&self,
length: usize
) -> Result<&[Value], ValueTypeError>

Convert the value to an array and return its elements.

pub fn as_i16_3(&self) -> [i16; 3]

Convert the value to an array of three i16s, panicking on failure.

pub fn try_as_i16_3(&self) -> Result<[i16; 3], ValueTypeError>

Convert the value to an array of three i16s.

pub fn as_f32_3(&self) -> [f32; 3]

Convert the value to an array of three f32s, panicking on failure.

pub fn try_as_f32_3(&self) -> Result<[f32; 3], ValueTypeError>

Convert the value to an array of three f32s.